requirements.txt
asgiref==3.2.7
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
oauthlib==3.1.0
protobuf==3.17.0
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.1.0
